feat: Add macOS platform support
Build System:
- Add macOS/ARM64 platform detection and configuration
- Configure vcpkg for macOS with proper triplet (arm64-osx)
- Build litehtml from source on macOS (no pre-built binaries)
- Handle macOS-specific library linking (framework paths, rpath)
- Add macOS entries to GDExtension configuration

Runtime Fixes:
- Fix SSL/SNI issue on macOS: enable proper SSL verification instead
  of disabling it (disabling VERIFYHOST also disables SNI, breaking
  modern servers like Cesium Ion)
- Add data URI handling for base64-encoded inline images (used by
  Cesium credit system logos)
- Add platform-specific includes for Marshalls class

Tested on macOS 15.1 (Sequoia) with Apple M1 Pro, Godot 4.5.1

+++ cesium_godot/Implementations/GodotPrepareRenderResources.cpp
@@ -145,7 +145,7 @@ void GodotPrepareRenderResources::free(Tile& tile, void* pLoadThreadResult, void
 	const auto& tileId = tile.getTileID();
 	size_t hash = std::visit(CesiumVariantHash{}, tileId);
 	auto* instance = static_cast<Cesium3DTile*>(pMainThreadResult);
-	this->m_tileset->call_deferred("free_tile", instance, hash);
+	this->m_tileset->call_deferred("free_tile", instance, static_cast<uint64_t>(hash));
 }
 
 void GodotPrepareRenderResources::attachRasterInMainThread(const Tile& tile, int32_t overlayTextureCoordinateID, const CesiumRasterOverlays::RasterOverlayTile& rasterTile, void* pMainThreadRendererResources, const glm::dvec2& translation, const glm::dvec2& scale)
